A more generous job buffer period
There are some problems with remote servers be a few seconds fast or
slow, this makes the job buffer period a bit more flexible. If a remote
client was fast, it was possible that a job would “succeed” but then an
alert would be sent because a job execution “missed” (even though it
was just early).