[CONFIG]  Updated sample configuration file to include necessary section
In order for the minicron client to connect to a hub server, there needs
to be a section defined in the configuration file for `client.server` that
specifies the values for `scheme`, `host`, `path` and `port`.  The use of these
values can be seen in `lib/minicron/cli/commands.rb`.  This commit updates the
sample configuration file to specify these needed values so that users do not
see a weird error message.