Import existing crontab jobs
Added a function to import the crontabs for a given host.
It will only import the jobs for the user registered to the host.

The function will get an array of jobs, save them to the database
and then will save their schedules to the database. It will then
return to the host page.